fixing profile and uer issues
