+++ server/controllers/profile.js
@@ -93,7 +93,8 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
         }
       });
 
-})
+});
+
 exports.findProfileById = asyncHandler(async (req, res, next) => {
 
      const profile = await Profile.findById(req.body.id);
@@ -115,3 +116,38 @@ exports.findAllProfiles = asyncHandler(async (req, res, next) => {
 
     res.status(200).json({ profiles: profiles });
 })
+
+const deletePhoto = async (user) => {
+  await cloudinary.uploader.destroy(user.profilePhoto.publicId);
+  user.profilePhoto = null;
+};
+
+exports.uploadProfilePhoto = asyncHandler(async (req, res, next) => {
+  const image = req.file;
+  const user = await Profile.findById(req.user.id);
+  if (!user) {
+    return res.sendStatus(404);
+  }
+  if (user.profilePhoto.url) {
+    await deletePhoto(user);
+  }
+  user.profilePhoto = {
+    url: image.path,
+    publicId: image.filename,
+  };
+  user.save();
+
+  res.sendStatus(200);
+});
+
+exports.deleteProfilePhoto = asyncHandler(async (req, res, next) => {
+  const user = await Profile.findById(req.user.id);
+  if (!user) {
+    return res.sendStatus(404);
+  }
+
+  await deletePhoto(user);
+  user.save();
+
+  res.sendStatus(200);
+});
