+++ server/db.js
@@ -9,14 +9,14 @@ const connectDB = async () => {
     useFindAndModify: false,
   });
   console.log(`MongoDB Connected: ${conn.connection.host}`.cyan.underline.bold);
-  if (process.env.NODE_ENV !== 'production' && mongoose.connection.readyState === 1) {
-    console.log('Seeding database...');
-    const user = await User.create({
-      username: 'demo',
-      email: 'demo@email.com',
-      password: 'password',
-    });
-  }
+  // if (process.env.NODE_ENV !== 'production' && mongoose.connection.readyState === 1) {
+  //   console.log('Seeding database...');
+  //   const user = await User.create({
+  //     username: 'demo',
+  //     email: 'demo@email.com',
+  //     password: 'password',
+  //   });
+  // }
 };
 
 module.exports = connectDB;
