+++ server/routes/profile.js
@@ -5,7 +5,11 @@ createProfile,
 updateProfile,
 findProfileById,
 findAllProfiles,
+uploadProfilePhoto,
+deleteProfilePhoto,
 } = require('../controllers/profile');
+const protect = require("../middleware/auth");
+const { multerUploads } = require("../utils/multer");
 
 router.route('/createProfile').post(createProfile);
 
@@ -14,5 +18,10 @@ router.route('/updateProfile').post(updateProfile);
 router.route('/findProfileById').post(findProfileById);
 
 router.route('/findAllProfiles').get(findAllProfiles);
+
+router
+  .route("/upload/profile-photo")
+  .post(protect, multerUploads.single("image"), uploadProfilePhoto);
+router.route("/profile-photo").delete(protect, deleteProfilePhoto);
 
 module.exports = router;
